Moved getValidatorProfileUrl method
